Feat: Add naver map api

+++ SW_Guide/app/src/main/java/com/sungshindev/sw_guide/data/Drink.java
@@ -5,8 +5,8 @@ public class Drink {
     String category;
     int drink_id;
     String image_path;
-    int lat;
-    int lon;
+    float lat;
+    float lon;
     String num;
     String recommend;
     String time;
@@ -28,8 +28,8 @@ public class Drink {
         return image_path;
     }
 
-    public int getLon(){return lon;}
-    public int getLat(){return lat;}
+    public float getLon(){return lon;}
+    public float getLat(){return lat;}
 
     public String getNum(){return num;}
     public String getRecommend(){return recommend;}
@@ -37,7 +37,7 @@ public class Drink {
     public String getCategory(){return category;}
 
     public Drink(int drink_id, String title,  String time, String num, String category,
-                 String image_path,int lon, int lat, String recommend){
+                 String image_path,float lon, float lat, String recommend){
         this.drink_id = drink_id;
         this.title = title;
         this.time = time;
